requirements.txt
fastapi[standard]
pydantic[email]
psycopg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv